create birthday reminder

+++ app/src/main/java/com/burtaev/application/ContactsDataSource.kt
@@ -2,5 +2,5 @@ package com.burtaev.application
 
 interface ContactsDataSource {
     fun getAllContact(): List<Contact>
-    fun getContactById(id: Long): Contact?
+    fun getContactById(id: Int): Contact?
 }
