+++ app/src/main/java/com/burtaev/application/OnContactClickedListener.kt
@@ -1,5 +1,5 @@
 package com.burtaev.application
 
 interface OnContactClickedListener {
-    fun showContactDetailsFragment(contactID: Long)
+    fun showContactDetailsFragment(contactID: Int)
 }
